Move import related logic into an import directory
