+++ src/originalBudgetTrackingApp/index.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import moment from 'moment';
-import { scrapeFinancialAccountsAndFetchTransactions } from '@/originalBudgetTrackingApp/importTransactions';
-import * as bankScraper from './bankScraper';
+import { scrapeFinancialAccountsAndFetchTransactions } from '@/originalBudgetTrackingApp/import/importTransactions';
+import * as bankScraper from './import/bankScraper';
 import * as ynab from './outputVendors/ynab/ynab';
 import { EnrichedTransaction } from './commonTypes';
 import * as configManager from './configManager/configManager';
